there are still errors in Crystal.unmerge!
I wrote a much cleaner implementation of Crystal.unmerge which is more
terse, faster and clearer. Unfortunately it still has some error which
causes certain reflections to be phase shifted 180 degrees from where
they ought to be. I will continue to work on this, but it seemed
reasonable to commit at this point, because the implementation is
already vastly improved.